+++ reactivestate/src/commonMain/kotlin/com/ensody/reactivestate/SimpleCoroutineLauncher.kt
@@ -15,15 +15,22 @@ public open class SimpleCoroutineLauncher(final override val launcherScope: Coro
     final override val isAnyLoading: LoadingStateTracker by lazy { LoadingStateTracker(launcherScope) }
     final override val generalLoading: MutableValueFlow<Int> by lazy { isAnyLoading.createLoadingState() }
 
+    protected open fun rawLaunch(
+        context: CoroutineContext,
+        start: CoroutineStart,
+        block: suspend CoroutineScope.() -> Unit,
+    ): Job =
+        launcherScope.launch(context = context, start = start, block = block)
+
     override fun launch(
         context: CoroutineContext,
         start: CoroutineStart,
         withLoading: MutableValueFlow<Int>?,
         onError: (suspend (Throwable) -> Unit)?,
-        block: suspend CoroutineScope.() -> Unit
+        block: suspend CoroutineScope.() -> Unit,
     ): Job =
-        launcherScope.launch(context = context, start = start) {
-            withErrorReporting({ onError?.invoke(it) ?: throw it }) {
+        rawLaunch(context = context, start = start) {
+            withErrorReporting({ onError?.invoke(it) ?: onError(it) }) {
                 try {
                     withLoading?.increment()
                     block()
@@ -32,4 +39,8 @@ public open class SimpleCoroutineLauncher(final override val launcherScope: Coro
                 }
             }
         }
+
+    public open fun onError(error: Throwable) {
+        throw error
+    }
 }