+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Non-breaking changes:
 
 * Added a multiplatform ViewModel `ReactiveState` (interface), `BaseReactiveState` (base class). This is actually a broader concept that can be used for any living object that can launch coroutines, automatically handles errors, triggers events, and tracks loading states.
 * Added a multiplatform `buildViewModel` extension function for creating such a `ViewModel` on an Activity and Fragment.
+* Improved automatic error catching for `autoRun` and `derived`.
 * Fixes for lifecycle observers: `onCreate`, `onCreateView`, `onCreateViewOnce`, `onDestroyView`, `onDestroyViewOnce`.
 
 Known limitations which will be solved with later releases:
@@ -28,6 +29,7 @@ Known limitations which will be solved with later releases:
 Changelog of preview releases:
 
 * 4.0.0-dev.2:
+  * Improved automatic error catching for `autoRun` and `derived`.
   * Fixes for lifecycle observers
 * 4.0.0-dev.1: This preview release comes without macOS/iOS builds. A port of the CI pipeline is in progress.